Update mavensmate-cli.coffee
atom.packages.resolvePackagePath('mavensmate') was failing to resolve for me resulting in an input to path.join that was 'undefined'. While 'mavensmate' should work with the fuzzy matching, it *does not* work on all installations. This PR changes the search string to the exact api name of the plugin, and resolves the problem.